// wanaku-router/ui/admin/src/models/promptReference.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * wanaku-router-backend API
 * OpenAPI spec version: 0.0.9
 */
import type { PromptMessage } from "./promptMessage";
import type { PromptArgument } from "./promptArgument";

export interface PromptReference {
  id?: string;
  name?: string;
  description?: string;
  messages?: PromptMessage[];
  arguments?: PromptArgument[];
  toolReferences?: string[];
  namespace?: string;
  configurationURI?: string;
}